Workaround to build the app on Xcode 6.1 and 6.1.1
When the project is built on Travis CI's Mac workers -- with Xcode 6.1
running on OS X 10.9 -- the Swift compiler complains about four
instances of optional unwrapping. This is caused by changes in
UITableViewCell's property definitions. Here we use this workaround so
that the project still gets built with the most recent SDK definition,
and it also gets built by Travis CI.